requirements.txt
streamlit
pdfplumber
pandas
openpyxl
bcrypt
python-dotenv
streamlit-cookies-manager
email-validator
openpyxl